requirements.txt
anyio==4.9.0
beautifulsoup4==4.13.4
certifi==2025.1.31
charset-normalizer==3.4.1
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
python-dotenv==1.1.0
python-telegram-bot==22.0
requests==2.32.3
schedule==1.2.2
sniffio==1.3.1
soupsieve==2.6
typing_extensions==4.13.2
urllib3==2.4.0
